이것저것 추가

+++ src/main/java/com/example/napoli/domain/repository/CarpoolRepository.java
@@ -4,6 +4,6 @@ package com.example.napoli.domain.repository;
 import com.example.napoli.domain.entity.Carpool;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarpoolRepository extends JpaRepository<Carpool, Integer> {
+public interface CarpoolRepository extends JpaRepository<Carpool, Long> {
 
 }
